Start desktop UI
